+++ mskpy/config.py
@@ -10,13 +10,24 @@ import configparser
 
 def _find_config():
     """Locate the config file."""
-    path = os.path.join(os.path.expanduser("~"), '.config', 'mskpy',
-                        'mskpy.cfg')
-    if not os.path.exists(path):
-        _create_config(path)
-    return path
+    fn = os.path.join(os.path.expanduser("~"), '.config', 'mskpy',
+                      'mskpy.cfg')
+    if not os.path.exists(fn):
+        _create_config(fn)
+
+    return fn
+
+_defaults = (
+    # (section, parameter, value)
+    ('ephem.core', 'kernel_path', os.path.sep.join([os.path.expanduser("~"), 'data', 'kernels'])),
+    ('calib', 'cohen_path', os.path.sep.join([os.path.expanduser("~"), 'data', 'mid-ir'])),
+    ('spex', 'spextool_path', os.path.sep.join([os.path.expanduser("~"), 'local', 'idl', 'irtf', 'Spextool'])),
+    ('irs', 'rogue_masks_path', os.path.sep.join([os.path.expanduser("~"), 'data', 'spitzer', 'irs', 'calibration', 'rogue-masks'])),
+    ('irs', 'spice_path', os.path.sep.join([os.path.expanduser("~"), 'local', 'spitzer', 'spice'])),
+)
 
 def _create_config(fn):
+    from configparser import DuplicateSectionError
     path = os.path.dirname(fn)
     d = path.split(os.path.sep)
     for i in range(len(d)):
@@ -27,20 +38,37 @@ def _create_config(fn):
             os.mkdir(x)
 
     home = os.path.expanduser("~")
-    config = configparser.RawConfigParser()
-    config.add_section('ephem.core')
-    config.set('ephem.core', 'kernel_path',
-               os.path.sep.join([home, 'data', 'kernels']))
-    config.add_section('calib')
-    config.set('calib', 'cohen_path',
-               os.path.sep.join([home, 'data', 'mid-ir']))
-    config.add_section('spex')
-    config.set('spex', 'spextool_path',
-               os.path.sep.join([home, 'local', 'idl', 'irtf', 'Spextool']))
+    config = _verify_config(configparser.RawConfigParser())
 
     with open(fn, 'w') as outf:
         config.write(outf)
 
+def _verify_config(c):
+    """Verifies that all sections are contained in `c`, else adds them."""
+    from configparser import DuplicateSectionError
+
+    updated = False
+    for section, parameter, value in _defaults:
+        try:
+            c.add_section(section)
+            updated = True
+        except DuplicateSectionError:
+            pass
+
+        if c.get(section, parameter, fallback=None) is None:
+            c.set(section, parameter, value)
+            updated = True
+
+    return c, updated
+
 config_file = _find_config()
+if not os.path.exists(config_file):
+    _create_config(config_file)
+
 config = configparser.RawConfigParser()
 config.read(config_file)
+c, updated = _verify_config(config)
+if updated:
+    config = c
+    with open(config_file, 'w') as outf:
+        config.write(outf)