+++ mskpy/instruments/spitzer.py
@@ -4,9 +4,16 @@
 spitzer --- Spitzer instruments.
 ================================
 
+   Functions
+   ---------
+   irsclean
+   irsclean_files
+   moving_wcs_fix
+
    Classes
    -------
    IRAC
+   IRS
 
 """
 
@@ -20,10 +27,84 @@ except ImportError:
 
 from .instrument import Instrument, Camera, LongSlitSpectrometer
 
-__all__ = ['IRAC']
+__all__ = ['irsclean', 'irsclean_files', 'IRAC', 'IRS']
+
+campaign2rogue = {
+    'IRSX002500': 'IRS1',
+    'IRSX002600': 'IRS2',
+    'IRSX002700': 'IRS3',
+    'IRSX002800': 'IRS4',
+    'IRSX002900': 'IRS5',
+    'IRSX003000': 'IRS6',
+    'IRSX003100': 'IRS7',
+    'IRSX003300': 'IRS8',
+    'IRSX003400': 'IRS9',
+    'IRSX003500': 'IRS10',
+    'IRSX003600': 'IRS11',
+    'IRSX003700': 'IRS12',
+    'IRSX003800': 'IRS13',
+    'IRSX003900': 'IRS14',
+    'IRSX004000': 'IRS15',
+    'IRSX004100': 'IRS16',
+    'IRSX004300': 'IRS17',
+    'IRSX004500': 'IRS18',
+    'IRSX004600': 'IRS19',
+    'IRSX004800': 'IRS20',
+    'IRSX005000': 'IRS21.1',
+    'IRSX007100': 'IRS21.2',
+    'IRSX006900': 'IRS21.3',
+    'IRSX007000': 'IRS21.4',
+    'IRSX005200': 'IRS22',
+    'IRSX005300': 'IRS23.1',
+    'IRSX007300': 'IRS23.2',
+    'IRSX005500': 'IRS24',
+    'IRSX005700': 'IRS25',
+    'IRSX005800': 'IRS26',
+    'IRSX006000': 'IRS27',
+    'IRSX006100': 'IRS28',
+    'IRSX006300': 'IRS29',
+    'IRSX006500': 'IRS30',
+    'IRSX006700': 'IRS31',
+    'IRSX006800': 'IRS32',
+    'IRSX007200': 'IRS33',
+    'IRSX007400': 'IRS34',
+    'IRSX007500': 'IRS35',
+    'IRSX007600': 'IRS36',
+    'IRSX007700': 'IRS37',
+    'IRSX007800': 'IRS38',
+    'IRSX008000': 'IRS39',
+    'IRSX008100': 'IRS40',
+    'IRSX008200': 'IRS41',
+    'IRSX008300': 'IRS42',
+    'IRSX008400': 'IRS43',
+    'IRSX009800': 'IRS44',
+    'IRSX009900': 'IRS45',
+    'IRSX010000': 'IRS46',
+    'IRSX010100': 'IRS47',
+    'IRSX008900': 'IRS48',
+    'IRSX010200': 'IRS49',
+    'IRSX010300': 'IRS50',
+    'IRSX010400': 'IRS51.1',
+    'IRSX011600': 'IRS51.2',
+    'IRSX011400': 'IRS52',
+    'IRSX009400': 'IRS53',
+    'IRSX009500': 'IRS54',
+    'IRSX010600': 'IRS55',
+    'IRSX010700': 'IRS56',
+    'IRSX010800': 'IRS57.1',
+    'IRSX011700': 'IRS57.2',
+    'IRSX010900': 'IRS58.1',
+    'IRSX011800': 'IRS58.2',
+    'IRSX011900': 'IRS58.3',
+    'IRSX011000': 'IRS59.1',
+    'IRSX012000': 'IRS59.2',
+    'IRSX011100': 'IRS60',
+    'IRSX012200': 'IRS61.1',
+    'IRSX011200': 'IRS61.2'
+}
 
 class IRAC(Camera):
-    """InfraRed Array Camera
+    """Spitzer's Infrared Array Camera
 
     Attributes
     ----------
@@ -126,6 +207,361 @@ class IRAC(Camera):
 #            K[i] = (davint(nu, _sf * ft * nu0[i] / nu, nu[0], nu[-1])
 #                    / davint(nu, ft * (nu0[i] / nu)**2, nu[0], nu[-1]))
 #        return K
+
+class IRS(Instrument):
+    """Spitzer's Infrared Spectrometer.
+
+    Attributes
+    ----------
+    module : The current IRS module: SL1, SL2, Blue, Red, etc.  SH, LH, SL3, LL3 not yet implemented.
+
+    Examples
+    --------
+
+    """
+
+    modes = ['sl1', 'sl2', 'll1', 'll2', 'blue', 'red']
+
+    def __init__(self):
+        self.sl2 = LongSlitSpectrometer(
+            6.37 * u.um,
+            [32, 128],
+            1.8 * u.arcsec,
+            2.0,
+            0.073 * u.um,
+            R=90,
+            location=Spitzer)
+        self.sl1 = LongSlitSpectrometer(
+            10.88 * u.um,
+            [32, 128],
+            1.8 * u.arcsec,
+            2.06,
+            0.12 * u.um,
+            R=90,
+            location=Spitzer)
+        self.ll2 = LongSlitSpectrometer(
+            17.59 * u.um,
+            [33, 128],
+            5.1 * u.arcsec,
+            2.1,
+            0.21 * u.um,
+            R=90,
+            location=Spitzer)
+        self.ll1 = LongSlitSpectrometer(
+            29.91 * u.um,
+            [33, 128],
+            5.1 * u.arcsec,
+            2.1,
+            0.35 * u.um,
+            R=85,
+            location=Spitzer)
+        self.blue = Camera(
+            15.8 * u.um,
+            [31, 44],
+            1.8 * u.arcsec,
+            location=Spitzer)
+        self.red = Camera(
+            22.3 * u.um,
+            [32, 43],
+            1.8 * u.arcsec,
+            location=Spitzer)
+
+        self._mode = 'sl1'
+
+    @property
+    def mode(self):
+        if self._mode in self.modes:
+            return self.__dict__[self._mode]
+        else:
+            raise KeyError("Invalid mode: {}".format(self._mode))
+
+    @mode.setter
+    def mode(self, m):
+        if m.lower() in self.modes:
+            self._mode = m.lower()
+        else:
+            raise KeyError("Invalid mode: {}".format(m.lower()))
+
+    def sed(self, *args, **kwargs):
+        """Spectral energy distribution of a target.
+
+        Parameters
+        ----------
+        *args
+        **kwargs
+          Arguments and keywords depend on the current IRS mode.
+
+        Returns
+        -------
+        sed : ndarray
+
+        """
+        return self.mode.sed(*args, **kwargs)
+
+    def lightcurve(self, *args, **kwargs):
+        """Secular lightcurve of a target.
+
+        Parameters
+        ----------
+        *args
+        **kwargs
+          Arguments and keywords depend on the current IRS mode.
+
+        Returns
+        -------
+        lc : astropy Table
+
+        """
+        return self.mode.lightcurve(*args, **kwargs)
+
+def irsclean(im, h, bmask=None, maskval=28672,
+             rmask=None, func=None, nan=True, sigma=None, box=3,
+             **fargs):
+    """Clean bad pixels from an IRS image.
+
+    Parameters
+    ----------
+    im : ndarray
+      The image.
+    h : dict-like
+      FITS header keywords from the original data file.
+    bmask : ndarray, optional
+      The SSC pipeline BMASK array.
+    maskval : int, optional
+      Bitmask value applied to BMASK array to generate a bad pixel
+      map.  These values will be removed from `bmask` and returned.
+    rmask : ndarray, optional
+      The rogue mask array.
+    func : function, optional
+      Use this function to clean the image: first argument is the
+      image to clean, the second is the mask (`True` for each bad
+      pixel).  The default is `image.fixpix`.
+    nan : bool, optional
+      Set to `True` to also clean any pixels set to NaN.
+    sigma : float, optional
+      Set to sigma clip the image using a filter of width `box` and
+      clipping at `sigma`-sigma outliers.
+    box : int, optional
+      The size of the filter for sigma clipping.
+    **fargs
+      Additional keyword arguments are pass to `func`.
+
+    Returns
+    -------
+    cleaned : ndarray
+      The cleaned data.
+    h : dict-like
+      The updated header.
+    new_mask : ndarray, optional
+      The cleaned mask.
+
+    """
+
+    import scipy.ndimage as nd
+    from ..image import fixpix
+
+    cleaner = fixpix if func is None else func
+
+    mask = np.zeros_like(im, bool)
+    if nan:
+        mask += ~np.isfinite(im)
+    if bmask is not None:
+        mask += (bmask & maskval).astype(bool)
+        new_mask = bmask & (32767 | maskval)
+    if rmask is not None:
+        mask += rmask.astype(bool)
+
+    if sigma is not None:
+        stdev = nd.generic_filter(im, np.std, size=box)
+        m = nd.median_filter(im, size=box)
+        mask += ((im - m) / stdev) > sigma
+    
+    h.add_history('Cleaned with mskpy.instruments.spitzer.irsclean.')
+    h.add_history('irsclean: function={}, arguments={}'.format(
+        str(cleaner), str(fargs)))
+    
+    cleaned = cleaner(im, mask, **fargs)
+    if bmask is None:
+        return cleaned, h
+    else:
+        return cleaned, h, new_mask
+
+def irsclean_files(files, outfiles, uncs=True, bmasks=True,
+                   maskval=16384, rmasks=True, func=None, nan=True,
+                   sigma=None, box=3, **fargs):
+    """Clean bad pixels from a list of IRS files.
+
+    For automatic rogue mask file name gleaning, this function
+    requires that irs.rogue_masks_path is set in mskpy.cfg to the
+    location of the rogue masks files from the Spitzer Science Center.
+
+    Parameters
+    ----------
+    files : array of strings
+      A list of image names to clean.
+    outfiles : array-like
+      Save cleaned images to these files.  Existing files will be
+      overwritten.  Uncertainty file names will be based on `outfiles`.
+    uncs : list or bool, optional
+      Also clean these uncertainty arrays.  Set to `True` and the
+      uncertainty filename will be guessed.  Otherwise set to `False`.
+    bmasks : ndarray, optional
+      The SSC pipeline BMASK arrays.  Same format as `uncs`.
+    maskval : int, optional
+      Bitmask value applied to BMASK array to generate bad pixel
+      maps.
+    rmasks : ndarray, optional
+      The rogue mask arrays.  Same format as `uncs`.  When `True` but
+      the IRS campaign is not in the `capaign2rogue` array (e.g., for
+      early release observations), then no rogue mask will be used.
+    func : function, optional
+      Use this function to clean the images: first argument is the
+      image to clean, the second is the mask (`True` for each bad
+      pixel).  The default is `image.fixpix`.
+    nan : bool, optional
+      Set to `True` to also clean any pixels set to NaN.
+    sigma : float, optional
+      Set to sigma clip the image using a filter of width `box` and
+      clipping at `sigma`-sigma outliers.
+    box : int, optional
+      The size of the filter for sigma clipping.
+    **fargs
+      Additional keyword arguments are pass to `func`.
+
+    """
+
+    from astropy.io import fits
+    
+    def file_generator(in_list, optional_list, replace_string):
+        for i in range(len(in_list)):
+            if optional_list is True:
+                f = in_list[i].replace('_bcd', replace_string)
+            elif np.iterable(optional_list):
+                f = optional_list[i]
+            else:
+                f = None
+
+            if f is None:
+                yield None, None
+            else:
+                yield f, fits.getdata(f)
+
+    def rmask_file_generator(in_list, optional_list):
+        import os.path
+        from ..config import config
+        path = config.get('irs', 'rogue_masks_path')
+        for i in range(len(in_list)):
+            if optional_list is True:
+                h = fits.getheader(in_list[i])
+                if h['CAMPAIGN'] in campaign2rogue:
+                    f = 'b{}_rmask_{}.fits'.format(
+                        h['CHNLNUM'], campaign2rogue[h['CAMPAIGN']])
+                    f = os.path.join((path, f))
+                else:
+                    f = None
+            elif np.iterable(optional_list):
+                f = optional_list[i]
+            else:
+                f = None
+
+            if f is None:
+                yield None, None,
+            else:
+                yield f, fits.getdata(f)
+
+    unc_files = file_generator(files, uncs, '_func')
+    bmask_files = file_generator(files, bmasks, '_bmask')
+    rmask_files = rmask_file_generator(files, rmasks)
+
+    for i in range(len(files)):
+        unc_file, unc = next(unc_files)
+        bmask_file, bmask = next(bmask_files)
+        rmask_file, rmask = next(rmask_files)
+
+        im, h = fits.getdata(files[i], header=True)
+        cleaned = irsclean(im, h, bmask=bmask, maskval=maskval,
+                           rmask=rmask, func=func, sigma=sigma,
+                           box=box, **fargs)
+        fits.writeto(outfiles[i], cleaned[0], cleaned[1], clobber=True)
+
+        if len(cleaned) == 3:
+            # bmask was updated, save it
+            bmask = cleaned[2]  # update array for use with unc cleaning
+            h_bmask = fits.getheader(bmask_file)
+            h_bmask.add_history('Updated with mskpy.instruments.spitzer.irsclean')
+            fits.writeto(bmask_file, bmask, h_bmask, clobber=True)
+        
+        if unc is not None:
+            if '_bcd' in outfiles[i]:
+                outf = outfiles[i].replace('_bcd', '_func')
+            elif outfiles[i].endswith('.fits'):
+                outf = outfiles[i].replace('.fits', '_func.fits')
+            else:
+                outf = outfiles[i] + '_func.fits'
+
+            h = fits.getheader(unc_file)
+
+            # do not use sigma clipping with unc array!
+            cleaned = irsclean(unc, h, bmask=bmask, maskval=maskval,
+                               rmask=rmask, func=func, sigma=None, **fargs)
+            
+            fits.writeto(outf, cleaned[0], cleaned[1], clobber=True)
+
+def moving_wcs_fix(files, ref=None):
+    """Correct IRS FITS WCS for the motion of the targeted moving object.
+
+    Parameters
+    ----------
+    files : array of strings
+      A list of files to update.  The files are updated in place.
+    ref : tuple
+      The "reference" RA and Dec of the target expressed as a tuple:
+      `(ra_ref, dec_ref)`.  This is usually the position of the moving
+      target at the start of the IRS observation.  The difference
+      between ra_ref, dec_ref and the RA_REF, DEC_REF in the FITS
+      headers is the motion of the target.  Set ref to `None` to use
+      RA_REF and DEC_REF from the first file in the file list as the
+      initial position.  [units: degrees]
+
+    """
+
+    from astropy.io import fits
+    from ..util import spherical_coord_rotate
+    
+    assert np.iterable(files), "files must be an array of file names"
+
+    ra_ref0, dec_ref0 = ref
+
+    for f in files:
+        im, h = fits.getdata(f, header=True)
+        ra_ref1 = h["RA_REF"]
+        dec_ref1 = h["DEC_REF"]
+
+        # I found CRVALx missing in some LH files
+        if h.get("CRVAL1") is not None:
+            crval1, crval2 = spherical_coord_rotate(
+                ra_ref1, dec_ref1, ra_ref0, dec_ref0,
+                h["CRVAL1"], h["CRVAL2"])
+        rarqst, decrqst = spherical_coord_rotate(
+            ra_ref1, dec_ref1, ra_ref0, dec_ref0,
+            h["RA_RQST"], h["DEC_RQST"])
+
+        raslt, decslt = spherical_coord_rotate(
+            ra_ref1, dec_ref1, ra_ref0, dec_ref0,
+            h["RA_SLT"], h["DEC_SLT"])
+
+        print("{} moved {:.3f} {:.3f}".format(f, (ra_ref1 - ra_ref0) * 3600.,
+                                              (dec_ref1 - dec_ref0) * 3600.))
+
+        if h.get("CRVAL1") is not None:
+            h["CRVAL1"] = crval1
+            h["CRVAL2"] = crval2
+        h["RA_RQST"] = rarqst
+        h["RA_SLT"] = raslt
+        h["DEC_RQST"] = decrqst
+        h["DEC_SLT"] = decslt
+        h.add_history("WCS updated for moving target motion with mskpy.instrum3ents.spitzer.moving_wcs_fix")
+        fits.update(f, im, h)
 
 # update module docstring
 from ..util import autodoc